--- src/game/server/core/components/Guilds/Manager/Members/Requests/GuildRequestData.cpp
@@ -1,3 +1,0 @@
-/* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
-/* If you are missing that file, acquire a complete release at teeworlds.com.                */
-#include "GuildRequestData.h"
